QCamera2: Fix the FD ROI reporting error in case of Bokeh
Read the preview dimensions inside updateConfigSettings function for
Bokeh mode as well. Preview dimensions are used to check if the translated
FD ROI is out of bounds.
Use the zoom to update shift used in the ROI translation as the shift is
now wrt sensor coordinates.
Use the low light, macro scene thresholds only when
FOVC_MAIN_CAM_FALLBACK_MECHANISM is enabled.

CRs-Fixed: 2022822

Change-Id: Ic012d8bd6fdbf67282e0fe2ea5a0d09c554e6a76