QCamera2: HAL: Stability issue in Dual Camera use case.
Issue:
In Dual camera on/off usecase, CAM_stMachine thread was
stuck in stop stream and waiting for deferredwork to be
finished.
Fix:
Currently sNextJobId is defined as static variable and is
shared through all QCamera2HardwareInterface instances.
sNextJobId should not be static because two independent
sessions are opened for dual camera and these two session
instances don't need to share sNextJobId.

Change-Id: I987bdec593e817710b498e5b7e3852707a18fc5e