QCamera2: Fix camera role and preview configuration in W+T.
Issue:
Currently Wide sensor is assumed to be Main and Tele as Aux.
But if the configuration is different, then the camera roles
and preview no display mode settings will go wrong.

Fix:
Assign roles and no display mode settings based on whether
wide is main camera or not.

CRs-Fixed: 2117488
Change-Id: I2d6904d287048afd2986c3b3c678b385974bf186